Replace np.int and numpy.int by int

+++ tslearn/tests/test_serialize_models.py
@@ -38,7 +38,7 @@ def clear_tmp():
 
 
 def test_hdftools():
-    dtypes = [numpy.int, numpy.int8, numpy.int16, numpy.int32, numpy.int64,
+    dtypes = [int, numpy.int8, numpy.int16, numpy.int32, numpy.int64,
               numpy.float, numpy.float32, numpy.float64]
 
     d = {}
